Refactor model modules: SpocoNet and "get_" methods
